backend/requirements.txt
botocore==1.25.4
jmespath==1.0.0
pynamodb==5.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
urllib3==1.26.9
